# wat: enforces signature order