This change does two things:
 * enforces field names are first or not present
   * Ex. (func $main) $main is the name of the func field
 * enforces the order requirements of the type grammar. (param)* result?
   * To reduce complexity, this logic is extracted to type_parser.

## Notes

### Unit test coverage is the same
There is still 100pct coverage except TODO and panics

### benchmark speed decrease and allocation increase

Extracting typeParser resulted in a performance decrease on the trivial benchmark because it initializes, but never needs a type parser. I think this is ok because in real life, %.wat files with anything non-trivial will end up with at least a few types.

decrease in speed
```
BenchmarkTextToBinaryExample/vs_wat.parseModule-16    	 1429618	       832.1 ns/op	     568 B/op	      23 allocs/op
BenchmarkTextToBinaryExample/vs_wat.parseModule-16    	 1293886	       977.1 ns/op	     664 B/op	      27 allocs/op
```

example input
```
(module $simple
	(import "" "hello" (func $hello))
	(start $hello)
)
```

Once we implement functions, we should re-do the benchmark anyway.

### Function pointer fake
The parser used by the lexer changes when fields change. It now also changes when we enter a common grammar (type use). To reduce stack depth and allocations, I'm updating moduleParser.tokenParser via a pointer to the moduleParser as it seems there's no function pointer in go.