@codefromthecrypt 2022-01-10
To check my assumptions about performance, I ran benchmarks against a file that has two types in it. The perf is basically the same, while both more strict enforcing grammar and decoupled code vs before.


before
```
BenchmarkTextToBinaryExample/vs_wat.parseModule
BenchmarkTextToBinaryExample/vs_wat.parseModule-16    	  433776	      2401 ns/op	    1200 B/op	      66 allocs/op
BenchmarkTextToBinaryExample/TextToBinary
BenchmarkTextToBinaryExample/TextToBinary-16          	  387208	      2809 ns/op	    1744 B/op	      77 allocs/op
```

Now
```
BenchmarkTextToBinaryExample/vs_wat.parseModule
BenchmarkTextToBinaryExample/vs_wat.parseModule-16    	  480280	      2411 ns/op	    1184 B/op	      65 allocs/op
BenchmarkTextToBinaryExample/TextToBinary
BenchmarkTextToBinaryExample/TextToBinary-16          	  409278	      2960 ns/op	    1728 B/op	      78 allocs/op
```

```
(module
	(import "wasi_snapshot_preview1" "path_open" (func $runtime.path_open (param i32 i32 i32 i32 i32 i64 i64 i32 i32) (result i32)))
	(import "wasi_snapshot_preview1" "fd_write" (func $runtime.fd_write (param i32 i32 i32 i32) (result i32)))
)
```